implementati vincoli (parte 2)

+++ src/main/java/it/addvalue/coverage/bean/PlanCalendarDetail.java
@@ -4,7 +4,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 @EqualsAndHashCode(of = "id")
@@ -12,8 +11,8 @@ public class PlanCalendarDetail implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	private Long       id;
-	private Long       idService;
-	private BigDecimal markerMultiplier;
+	private Long    id;
+	private Long    idService;
+	private Integer markerMultiplier;
 
 }
